lib/lintFiles.js: Accept news as a changelog file
There are a number of projects that ship with a NEWS file, a summary of
changes between release versions, and do not have a ChangeLog file, per
se. The rationale for including NEWS as an acceptable change log is that
it is shorter and easier to follow than a full changelog, thus, useful
for both users and contributors.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>